Update MSI-X doc to permit NDIS to rebalance resources at any time
NDIS docs originally claimed that NDIS only calls MiniportFilterResourceRequirements while the miniport is halted.  However, this is not how WDM actually works -- WDM permits the rebalance IRP to come in at any time.  Experimentally we see IRP_MN_FILTER_RESOURCE_REQUIREMENTS come in while the PNP device is in a query-stopped state, not a stopped state.