Clarify what "output" variable stands for
In existing example: 
var output = ctl.ExecuteCommand("u");
It's not clear what type "output" represents. Is it a string blob? Is it some internal log type?
It turns out that it's actually an array of string lines. Changing "output" variable name to "outputLines" could make it more clear that it's an array.
Proposed change:
var outputLines = ctl.ExecuteCommand("u");