Grammer: possessive form
Updated the sentence: "COSA[…]is the new format Mobile Operators (MOs) 'use' in Windows 10, version 1703 and later to provision Windows devices for mobile broadband." Changed the "use" to "used".

This is not a past tense. It is a possessive form. Something is USING Windows 10 to accomplish a task, in this case to provision devices for mobile broadband.

Let's give a couple examples. "This car is used by Emily for her daily commutes". This sentence is still in present tense. To say "This car is use by Emily for her daily commutes" would be grammeritacally incorrect.

Another example: "Visual Studio 2017 is used for developing UWP apps". This is a present tense, but, again, note the "used" is correct here. If it were past tense, it would be something like this: "Visual Studio 6.0 was used for C++ and VB6 development in the 90s". Note that the ver "to be" is either "is" or "was", depending on whether the sentence is past or present and the verb "to use" is alwasy written as "used" in both cases.