Deleted repetition; minor layout changes
First, I deleted repetitions of "the Boot.ini file". We need to establish that "Boot.ini" is a file only once. Once it is done, we can just say "Boot.ini" instead of "the Boot.ini file". Suprisingly, this principle was already observed in regards to "ntldr". Also, I deleted an unused visible but empty anchor.