Update ndis-status-wwan-pin-info.md
The suggested change is from Cellcore feature team. The original statement has conflicts with others.
It indicates "all other members are ignored"; however, it should depend on the situations.
For example, the following statement says:
"When a PIN mode is changed from disabled to enabled, the PIN state should be WwanPinStateNone."

Without the change, one may not know whether PIN state should be ignored or be set to  WwanPinStateNone.