More samples and explanation of request processing
When I first read this page, it was clear what was needed to do for set up of the objects all the way to endpoints and queues.  However, it was not evident what callbacks needed to be processed on the endpoint-associated queues. I'm hoping this update will help the next person.