Delete a new "Note" section and add a sentence
Delete a new "Note" section and add a sentence based on Brandon Jiang's advice:
Hi, Akihisa, instead of adding a new “Note” section, can you just add a sentence at end of the following paragraph? 

*NdisDeviceType
The type of the device. The default value is zero, which indicates a standard networking device that connects to a network. Set *NdisDeviceType to NDIS_DEVICE_TYPE_ENDPOINT (1) if this device is an endpoint device and is not a true network interface that connects to a network. For example, you must specify NDIS_DEVICE_TYPE_ENDPOINT for devices such as smart phones that use a networking infrastructure to communicate to the local computer system but do not provide connectivity to an external network; On the other hand, you must *not* set this keyword for virtual adapter such like VPN interface as they actually provides connectivity to an external network.