Can't Assume IRQL PASSIVE_LEVEL
Please don't encourage people to assume the IRQL will be PASSIVE_LEVEL, as the parenthetical in paragraph at line 239 does now.  Either remove the parenthetical entirely (this is the best idea) or at the warning  that the driver code must never RELY on the IRQL being PASSIVE_LEVEL in this case.  The only time a driver has the "right" to assume the IRQL is PASSIVE_LEVEL in an EvtIo Event Processing Callback is when the driver has established the appropriate ExecutionLevel constraint.  No other time.  Anything else requires people to rely upon an artifact of implementation (that the user-mode caller's IRQL is passed-through by the Framework), which is a mistake.