Update other-wdk-downloads.md
Revised to indicate that the reader must install the SDK seprately for 1607, in addition to 1703, since for both verions of Windows, they should install Visual Studio 2015, which did not include the SDK.

I also commented out "Remote Debugging client for Windows 8.1" in "Step 2: Install the DWK". Don Marshall and Cymon Kilmer both agreed it should go.