adding MSDL service status + contact info (#1449)
The goal here is to have a "status box" on this page to display to customers whether the public symbols server is up or not + an official avenue to report issues, which we currently don't have. The status will be edited as part of any livesite work if issues come in so will require a manual edit from our on-call folks. Historically this has only happened once every two months or so. The next step is to see if we can do any automated azure monitor piping so the status can be automatically updated. I just realized I didn't edit this with a corporate GitHub account...