C99651: Escaping exclamation mark to avoid broken content in localized pages
Localization team has reported source content issue that causes localized version to have broken/different format compared to en-us version.  
Description: Description of the source issue:  The transformer recognizes the text as an alternative comment for an image because of the exclamation mark. Please consider adding a backslash to escape the exclamation mark.
@tedhudek