Fix tracelogging-examples.md
The existing example was apparently copied from an old Microsoft-internal sharepoint page that had been written very early in the development of TraceLoggingProvider.h. The sharepoint page was intended as a technical demonstration, not as a public sample. The TraceLoggingProvider.h header was substantially modified before it was released, so this sample doesn't even compile. In addition, the sample refers to several other headers that are not provided and are not public.

I've made some quick fixes to this sample so that it can compile and doesn't mislead users.