Adding Sysinternal's ProcDump + Removing userdump
I am adding Sysinternal's ProcDump and removing ADPlus and Userdump.exe. 

ProcDump's feature set has made it the primary tool for Microsoft Support when collecting dumps for real-world customer issues.
ADPlus has not been used by support for a long time and Userdump is from the OEM support tools which can't be found anymore. The links in the referenced KB for Userdump are all dead too.