Revise the OS versions
Based on Sandy's advice, changed this misleading text, “An attestation signed driver works on Windows 10 and Windows Server 2016 and later. It does not work on earlier versions of Windows, such as Windows 8.1 and Windows 7.”

To this text “An attestation signed driver works on Windows 10 It does not work on earlier versions of Windows, such as Windows 8.1 and Windows 7, and is not supported for Windows Server 2016 and later.  See this link for more information on the support policy, https://support.microsoft.com/en-us/help/4519013/support-policy-3rd-party-kernel-level-attestation-signed-software.”