Remove misleading statement.
Removes the line "Possible violations include ... A resource that
is released out of sequence".

Release a resource is always (effectively) non blocking and so cannot
cause a deadlock.

I don't think the author meant to mean that dropping a resource could
cause a deadlock, but it is a common enough misconception that we need
to avoid phrases which might make it more common.

I constant run with verifier on and often have to release in a
non-acquired order and I have never hit this verifier error, I therefore
assume is has been removed from the verifier (and I guess that if it
hasn't been it should be).