Note to not use Deadlock Detection with Kernel synch fuzzing.
You get a false warning Typically

Thread A - owns the spinlock and is releasing it but verifier
is stalling the release

Thread B - owns the resource and is queuing for the spinlock

Thread C - reports a deadlock on the resource.

It is obviously bogus that a spinlock and an eresource can deadlock
each other, but that is another story.

The bottom line is that these two options are mutually incomaptible