Fix formatting in advanced-symsrv-use.md (#2930)
Not sure what happened here, it looks like the garbled result of multiple rounds of asterisk escaping. I've tried to correct and simplify it as much as possible. I kept the partial emphasis in the cache\*localsymbolcache example so that it's clear that `localsymbolcache` is a placeholder for a path and not a literal string.