Update documentation to note that 3A ROI coordinates are window-dependent (#3155)
* Update documentation to note that 3A ROI coordinates are window-dependent

When the Platform DMFT is used for face-based 3A, the DMFT will calculate the face coordinates based on the current zoom window of the image. If the camera has been zoomed in (e.g., due to the PTZ or Digital Window controls), the camera is responsible for mapping the coordinates we provide back to the "real" FoV - essentially undoing the zoom. Updating the documentation to make this specific.

* Clarify mapping between full and modified FoV