Additional improvements for Arm/RISC-V platforms (#3460)
* SPCR: allow faster baud rates.

On some SoC platforms, the UART can run significantly faster than 115200.

Propose 230400, 460800, 500000, 576000, 921600, 1000000, 1152000,
1500000, 2000000, 2500000, 3000000, 3500000 and 4000000 to be added to
the list of supported baud rates. This matches the baud rates > 115200
that are supported by the Linux kernel.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* DBG2: clarify language for NamespaceString[]

Existing language could be interpreted as any string containing a period.

Also, NUL characters - not null characters.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* SPCR: AML Namespace string

The SPCR can describe non-PCIe UARTs. To facilitate hand-off between
an OS kernel's early console support code and the proper serial
hardware driver, the SPCR specification could be amended with
additional fields, similar to the ones present in DBG2, to link to the
AML description of the device.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

---------

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>