restore some numbers and a "shall"
A recent language clean-up introduced some stylistic changes that diverged the Microsoft-defined Bluetooth HCI extensions from the format used in the Bluetooth SIG specifications that we're modelling off of. Additionally, the change removed a "shall" from the text which, based on the language standards specified by the Bluetooth SIG, removes our ability to enforce a requirement from that text.

This change proposes to revert some of the changes.