RISC-V support and assorted improvements to SPCR (#3463)
* DBG2: define 0x0015 as the SBI console serial subtype

When 0x0015 is reported, the OS is to follow the SBI specification,
probing the actual available mechanism for console output. This may
ether be the DBCN extension or the legacy console EIDs.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* SPCR: add RISC-V InterruptType

Define Bit[4] as a GSI for any RISC-V interrupt controller (PLIC, APLIC).

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* SPCR: allow InterruptType to be 0.

Quirks side, certain interfaces like the RISC-v SBI console
do not support interrupt-based notifications.

Thus also clarify the GSI field to be valid if at least 1
bit in Bit[1:7] of InterruptType is set.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* SPCR: allow faster baud rates.

On some SoC platforms, the UART can run significantly faster than 115200.

Introduce a new 32-bit field, Baud Rate Extended, which is to be honored
when the 8-bit Baud Rate field is reported as 255.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* DBG2: clarify language for NamespaceString[]

Existing language could be interpreted as any string containing a period.

Also, NUL characters - not null characters.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* SPCR: AML Namespace string

The SPCR can describe non-PCIe UARTs. To facilitate hand-off between
an OS kernel's early console support code and the proper serial
hardware driver, the SPCR specification could be amended with
additional fields, similar to the ones present in DBG2, to link to the
AML description of the device.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* Tweak baudrate fields and add descriptive text; add extra paragraph on 16550; update rev history

* Update SPCR rev to 4; fix field offsets for NS string; tweak new baud rate fields

* Tweaks

* Typo

* Consistent capitalization of NamespaceString

* Fix indentation level

* spcr: amend revision comment

...to also cover the enhancements to interrupt type field.

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>

* Tweak language in 16550 note

* Tweak SPCR language

---------

Signed-off-by: Andrei Warkentin <andrei.warkentin@intel.com>
Co-authored-by: Brian Nezvadovitz <brinez@microsoft.com>