Add warning about "Release for Mod Write" being called as a special kernel APC
Been bitten by this twice now.

This callbacks comes in as a special kernel mode APC and so can interupt another thread even if they have
called EnterCriticalRegion.

If such a thread (for instance) holds an ERESOURCE shared and the callback tries to get it Ex you will deadlock.

The only way I have foind to make this work is to ensure that this callback only releases locks (not hard, but
bears remembering).

This PR probably requires extensive word-smithing.