Remove "/ac" from the signtool command
The "/ac" and its accompanying argument is a remnant from the "old days" before EV Certs.  EV Certs can't be stored in a local path anymore (they're on a token or other hardware device)... so /ac doesn't apply anymore.  The command works perfectly with /ac <path> removed (we've done this, multiple times).