Update code-signing-reqs.md, remove Entrust
Replace the link to Entrust with Sectigo, since Entrust stopped issuing certificates and refers to Sectigo. See https://store.entrust.com/code-signing-solutions

A point of clarification here though, is this list exhaustive? It is not made clear if these are the *only* valid certificate authorities for EV Certificates, or if you can use any authority in the MS Trusted Root Program, a bit of clarification here would help.