Add Help Page for Open Dump Errors
One of the most frequent negative user feedback we get is on failures to
open dump files in WinDbg. The error presented to the user is not always
clear on why the dump failed to open and what the user can do about it.

This change adds resolution steps for some common errors when opening
dumps.