Update windbg data model explorer docs (#2833)
* Update windbg data model explorer docs

This change updates text and images of the data model window with the
latest UI changes. Specific updates are omission of the change query
button and window in favor of newer query text box for editing queries.
Also, data model windows can be given a custom name so users can label
them as desired to track the various windows, as needed.

* Address Feedback

---------

Co-authored-by: John Schultz <joschult@microsoft.com>